feat: add mailbox list, import (append/overwrite/dedup), export, and delete API
Also fixes AuditLog BigInteger id to Integer for SQLite compatibility.

+++ backend/app/models/audit_log.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from sqlalchemy import String, Integer, DateTime, BigInteger, Text
+from sqlalchemy import String, Integer, DateTime, Text
 from sqlalchemy import JSON
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -10,7 +10,7 @@ from app.database import Base
 class AuditLog(Base):
     __tablename__ = "audit_logs"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     action: Mapped[str] = mapped_column(String(50), nullable=False, index=True)
     target_count: Mapped[int] = mapped_column(Integer, nullable=False)
     detail: Mapped[dict] = mapped_column(JSON, default=dict)
